app/pdf-to-qti/requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# AI Providers
google-genai>=0.3.0
openai>=1.0.0
boto3>=1.34.0

# CLI
click>=8.0.0

# Image processing (for validation with resizing)
Pillow>=10.0.0

# Optional: for local XSD validation
# lxml>=4.9.0

